pyproject.toml
[project]
name = "risk-mas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cvxpy>=1.7.5",
    "fastapi>=0.128.0",
    "jsonschema>=4.23.0",
    "langchain>=1.2.3",
    "langchain-google-genai>=4.1.3",
    "langchain-openai>=0.3.0",
    "langgraph>=1.0.5",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "tushare>=1.4.24",
    "uvicorn>=0.40.0",
]
